+++ src/core/VSCodeSetting.ts
@@ -1,86 +1,29 @@
 import * as async from "async";
-import * as fs from "fs";
+import * as fs from "fs-extra";
 import * as junk from "junk";
 import * as path from "path";
 import * as vscode from "vscode";
 
-import { CONFIGURATION_KEY, CONFIGURATION_POKA_YOKE_THRESHOLD, SETTINGS_UPLOAD_EXCLUDE } from "../constants";
+import { CONFIGURATION_KEY, CONFIGURATION_POKA_YOKE_THRESHOLD, SETTINGS_UPLOAD_EXCLUDE } from "../common/constants";
+import * as GitHubTypes from "../common/GitHubTypes";
+import { IExtension, ISetting, ISyncedItem, SettingTypes } from "../common/types";
 import { diff } from "../utils/diffHelper";
 import { excludeSettings, mergeSettings, parse } from "../utils/jsonHelper";
 import Environment from "./Environment";
-import Extension, { IExtension, ISyncStatus } from "./Extension";
-import * as GitHubTypes from "./GitHubTypes";
+import Extension from "./Extension";
 import * as Toast from "./Toast";
 
 /**
- * Represent the type of VSCode configs.
+ * `VSCode Settings` wrapper.
  */
-export enum ConfigTypes
+export default class VSCodeSetting
 {
-    /**
-     * Files.
-     */
-    File,
+    private static _instance: VSCodeSetting;
 
     /**
-     * Folders.
+     * Suffix of remote mac files.
      */
-    Folder
-}
-
-/**
- * Represent a VSCode config.
- */
-export interface IConfig
-{
-    /**
-     * Config file name.
-     */
-    name: string;
-
-    /**
-     * Config file path in local.
-     */
-    path: string;
-
-    /**
-     * Config file name in remote.
-     */
-    remote: string;
-
-    /**
-     * Config file content.
-     */
-    content?: string;
-}
-
-/**
- * Represent the options of [getConfigs](#Config.getConfigs).
- */
-export interface IGetConfigOptions
-{
-    /**
-     * Whether to load the full list of VSCode settings. Defaults to `false`.
-     */
-    full?: boolean;
-
-    /**
-     * Whether to load the content of VSCode settings files. Defaults to `false`.
-     */
-    load?: boolean;
-
-    /**
-     * Whether to show the progress indicator. Defaults to `false`.
-     */
-    showIndicator?: boolean;
-}
-
-/**
- * VSCode configs wrapper.
- */
-export default class Config
-{
-    private static _instance: Config;
+    private static readonly MAC_SUFFIX: string = "-mac";
 
     /**
      * Prefix of remote snippet files.
@@ -97,19 +40,19 @@ export default class Config
     }
 
     /**
-     * Create an instance of singleton class `Config`.
+     * Create an instance of singleton class `VSCodeSetting`.
      */
-    public static create(context: vscode.ExtensionContext): Config
+    public static create(context: vscode.ExtensionContext): VSCodeSetting
     {
-        if (!Config._instance)
+        if (!VSCodeSetting._instance)
         {
-            Config._instance = new Config(context);
+            VSCodeSetting._instance = new VSCodeSetting(context);
         }
-        return Config._instance;
+        return VSCodeSetting._instance;
     }
 
     /**
-     * Get VSCode settings list (will be uploaded or downloaded, anyway).
+     * Get `VSCode Settings` (will be uploaded or downloaded, anyway).
      * For example:
      *    [
      *        {
@@ -120,12 +63,15 @@ export default class Config
      *        },
      *        ...
      *    ]
+     *
+     * @param {boolean} [loadFileContent=false] Whether to load the content of `VSCode Settings` files. Defaults to `false`.
+     * @param {boolean} [showIndicator=false] Whether to show the progress indicator. Defaults to `false`.
      */
-    public getConfigs({ full = false, load = false, showIndicator = false }: IGetConfigOptions = {}): Promise<IConfig[]>
+    public getSettings(loadFileContent: boolean = false, showIndicator: boolean = false): Promise<ISetting[]>
     {
         return new Promise((resolve) =>
         {
-            function resolveWrap(value: IConfig[])
+            function resolveWrap(value: ISetting[])
             {
                 if (showIndicator)
                 {
@@ -139,75 +85,56 @@ export default class Config
                 Toast.showSpinner("Syncing: Gathering local settings.");
             }
 
-            // The item order is very important to ensure that the small files are synced first.
+            // Note that this is an ordered list, to ensure that the smaller files (such as `settings.json`, `keybindings.json`) are synced first.
             // Thus, the extensions will be the last one to sync.
-            const list: Array<{ name: string, type: ConfigTypes }> = [
-                { name: "locale", type: ConfigTypes.File },
-                { name: "snippets", type: ConfigTypes.Folder },
-                { name: "extensions", type: ConfigTypes.File }
+            const settingsList = [
+                SettingTypes.Settings,
+                SettingTypes.Keybindings,
+                SettingTypes.Locale,
+                SettingTypes.Snippets,
+                SettingTypes.Extensions
             ];
 
-            if (full)
-            {
-                list.unshift(
-                    { name: "settings-mac", type: ConfigTypes.File },
-                    { name: "settings", type: ConfigTypes.File },
-                    { name: "keybindings-mac", type: ConfigTypes.File },
-                    { name: "keybindings", type: ConfigTypes.File }
-                );
-            }
-            else
-            {
-                list.unshift(
-                    this._env.isMac ? { name: "settings-mac", type: ConfigTypes.File }
-                        : { name: "settings", type: ConfigTypes.File },
-                    this._env.isMac ? { name: "keybindings-mac", type: ConfigTypes.File }
-                        : { name: "keybindings", type: ConfigTypes.File }
-                );
-            }
-
-            let temp: IConfig[];
+            let tempSettings: ISetting[];
             let localFilename: string;
-            const results: IConfig[] = [];
+            let remoteFilename: string;
+            const results: ISetting[] = [];
             const errorFiles: string[] = [];
             async.eachSeries(
-                list,
-                (item, done) =>
+                settingsList,
+                (type, done) =>
                 {
-                    if (item.type === ConfigTypes.Folder)
+                    if (type === SettingTypes.Snippets)
                     {
                         // Attention: Snippets may be empty.
-                        temp = this._getSnippets(this._env.snippetsPath);
+                        tempSettings = this._getSnippets(this._env.snippetsPath);
                     }
                     else
                     {
-                        localFilename = `${item.name}.json`;
-                        if (item.name.includes("settings"))
+                        localFilename = `${type}.json`;
+                        remoteFilename = localFilename;
+                        if (type === SettingTypes.Settings || type === SettingTypes.Keybindings)
                         {
-                            localFilename = "settings.json";
-                        }
-                        else if (item.name.includes("keybindings"))
-                        {
-                            localFilename = "keybindings.json";
+                            // TODO: Merge settings files into one: The MAC_SUFFIX should be removed in the next release.
+                            remoteFilename = this._env.isMac
+                                ? `${type}${VSCodeSetting.MAC_SUFFIX}.json`
+                                : `${type}.json`;
                         }
 
-                        temp = [
+                        tempSettings = [
                             {
-                                name: item.name,
-                                path: path.join(
-                                    this._env.codeUserPath,
-                                    localFilename
-                                ),
-                                remote: `${item.name}.json`
+                                filepath: path.join(this._env.codeUserPath, localFilename),
+                                remoteFilename,
+                                type
                             }
                         ];
                     }
 
-                    if (load)
+                    if (loadFileContent)
                     {
-                        this._loadContent(temp).then((values: IConfig[]) =>
+                        this._loadContent(tempSettings).then((values: ISetting[]) =>
                         {
-                            values.forEach((value: IConfig) =>
+                            values.forEach((value: ISetting) =>
                             {
                                 // Check for success.
                                 if (value.content)
@@ -216,7 +143,7 @@ export default class Config
                                 }
                                 else
                                 {
-                                    errorFiles.push(value.remote);
+                                    errorFiles.push(value.remoteFilename);
                                 }
                             });
                             done();
@@ -224,7 +151,7 @@ export default class Config
                     }
                     else
                     {
-                        results.push(...temp);
+                        results.push(...tempSettings);
                         done();
                     }
                 },
@@ -241,20 +168,20 @@ export default class Config
     }
 
     /**
-     * Save VSCode settings to files.
-     * @param files VSCode Configs from Gist.
+     * Save `VSCode Settings` to files.
+     * @param files `VSCode Settings` from GitHub Gist.
      * @param showIndicator Whether to show the progress indicator. Defaults to `false`.
      */
-    saveConfigs(files: GitHubTypes.IGistFiles, showIndicator: boolean = false): Promise<{
-        updated: ISyncStatus[],
-        removed: ISyncStatus[]
+    saveSettings(files: GitHubTypes.IGistFiles, showIndicator: boolean = false): Promise<{
+        updated: ISyncedItem[],
+        removed: ISyncedItem[]
     }>
     {
         return new Promise((resolve, reject) =>
         {
             function resolveWrap(value: {
-                updated: ISyncStatus[],
-                removed: ISyncStatus[]
+                updated: ISyncedItem[],
+                removed: ISyncedItem[]
             })
             {
                 if (showIndicator)
@@ -280,44 +207,62 @@ export default class Config
 
             if (files)
             {
-                let extensionsFile: IConfig;
-                const saveFiles: IConfig[] = [];
-                const removeFiles: IConfig[] = [];
+                let extensionsSetting: ISetting;
+                const settingsToSave: ISetting[] = [];
+                const settingsToRemove: ISetting[] = [];
                 const existsFileKeys: string[] = [];
-                this.getConfigs().then((configs: IConfig[]) =>
+                this.getSettings().then((settings: ISetting[]) =>
                 {
-                    let file: GitHubTypes.IGistFile;
-                    for (const config of configs)
+                    let gistFile: GitHubTypes.IGistFile;
+                    for (const setting of settings)
                     {
-                        file = files[config.remote];
-                        if (file)
+                        gistFile = files[setting.remoteFilename];
+                        if (gistFile)
                         {
                             // File exists in remote and local, sync it.
-                            if (config.name === "extensions")
+                            if (setting.type === SettingTypes.Extensions)
                             {
                                 // Temp extensions file.
-                                extensionsFile = {
-                                    ...config,
-                                    content: file.content
+                                extensionsSetting = {
+                                    ...setting,
+                                    content: gistFile.content
                                 };
                             }
                             else
                             {
                                 // Temp other file.
-                                saveFiles.push({
-                                    ...config,
-                                    content: file.content
+                                settingsToSave.push({
+                                    ...setting,
+                                    content: gistFile.content
                                 });
                             }
-                            existsFileKeys.push(config.remote);
+                            existsFileKeys.push(setting.remoteFilename);
                         }
                         else
                         {
+                            // TODO: Merge settings files into one: Should be updated in the next release.
+                            if (setting.type === SettingTypes.Settings)
+                            {
+                                const fallbackSettingsRemoteFilename = this._env.isMac
+                                    ? `${setting.type}.json`
+                                    : `${setting.type}${VSCodeSetting.MAC_SUFFIX}.json`;
+                                const fallbackSettings = files[fallbackSettingsRemoteFilename];
+                                if (fallbackSettings)
+                                {
+                                    // Temp fallbackSettings file.
+                                    settingsToSave.push({
+                                        ...setting,
+                                        content: fallbackSettings.content
+                                    });
+                                    existsFileKeys.push(setting.remoteFilename);
+                                }
+                            }
+
                             // File exists in remote, but not exists in local.
                             // Delete if it's a snippet file.
-                            if (config.remote.startsWith(Config.SNIPPET_PREFIX))
+                            if (setting.type === SettingTypes.Snippets)
                             {
-                                removeFiles.push(config);
+                                settingsToRemove.push(setting);
                             }
                         }
                     }
@@ -327,17 +272,18 @@ export default class Config
                     {
                         if (existsFileKeys.indexOf(key) === -1)
                         {
-                            file = files[key];
-                            if (file.filename.startsWith(Config.SNIPPET_PREFIX))
+                            gistFile = files[key];
+                            if (gistFile.filename.startsWith(VSCodeSetting.SNIPPET_PREFIX))
                             {
-                                filename = file.filename.slice(Config.SNIPPET_PREFIX.length);
+                                // Snippets.
+                                filename = gistFile.filename.slice(VSCodeSetting.SNIPPET_PREFIX.length);
                                 if (filename)
                                 {
-                                    saveFiles.push({
-                                        content: file.content,
-                                        name: filename,
-                                        path: this._env.getSnippetFilePath(filename),
-                                        remote: file.filename
+                                    settingsToSave.push({
+                                        content: gistFile.content,
+                                        filepath: this._env.getSnippetFilePath(filename),
+                                        remoteFilename: gistFile.filename,
+                                        type: SettingTypes.Snippets
                                     });
                                 }
                             }
@@ -349,31 +295,31 @@ export default class Config
                     }
 
                     // Put extensions file to the last.
-                    if (extensionsFile)
+                    if (extensionsSetting)
                     {
-                        saveFiles.push(extensionsFile);
+                        settingsToSave.push(extensionsSetting);
                     }
 
                     // poka-yoke - check if there have been two much changes (more than 10 changes) since the last downloading.
-                    this._checkIfContinue(configs, saveFiles, removeFiles).then((value) =>
+                    this._shouldContinue(settings, settingsToSave, settingsToRemove).then((value) =>
                     {
                         if (value)
                         {
-                            const syncedFiles: {
-                                updated: ISyncStatus[],
-                                removed: ISyncStatus[]
+                            const syncedItems: {
+                                updated: ISyncedItem[],
+                                removed: ISyncedItem[]
                             } = { updated: [], removed: [] };
                             async.eachSeries(
-                                saveFiles,
-                                (item: IConfig, done: async.ErrorCallback<Error>) =>
+                                settingsToSave,
+                                (setting: ISetting, done: async.ErrorCallback<Error>) =>
                                 {
-                                    this._saveItemContent(item).then((saved) =>
+                                    this._saveSetting(setting).then((saved) =>
                                     {
-                                        syncedFiles.updated.push(saved);
+                                        syncedItems.updated.push(saved);
                                         done();
                                     }).catch((err) =>
                                     {
-                                        done(new Error(`Cannot save file: ${item.remote} : ${err.message}`));
+                                        done(new Error(`Cannot save file: ${setting.remoteFilename} : ${err.message}`));
                                     });
                                 },
                                 (err) =>
@@ -384,10 +330,10 @@ export default class Config
                                     }
                                     else
                                     {
-                                        this.removeConfigs(removeFiles).then((removed) =>
+                                        this.removeSettings(settingsToRemove).then((removed) =>
                                         {
-                                            syncedFiles.removed = removed;
-                                            resolveWrap(syncedFiles);
+                                            syncedItems.removed = removed;
+                                            resolveWrap(syncedItems);
                                         }).catch(rejectWrap);
                                     }
                                 }
@@ -408,29 +354,25 @@ export default class Config
     }
 
     /**
-     * Delete the physical files.
-     * @param files Files list.
+     * Delete the physical files corresponding to the `VSCode Settings`.
+     * @param settings `VSCode Settings`.
      */
-    removeConfigs(files: IConfig[]): Promise<ISyncStatus[]>
+    removeSettings(settings: ISetting[]): Promise<ISyncedItem[]>
     {
         return new Promise((resolve, reject) =>
         {
-            const removed: ISyncStatus[] = [];
+            const removed: ISyncedItem[] = [];
             async.eachSeries(
-                files,
+                settings,
                 (item, done) =>
                 {
-                    fs.unlink(item.path, (err) =>
+                    fs.remove(item.filepath).then(() =>
                     {
-                        if (err)
-                        {
-                            done(new Error(`Cannot save file: ${item.remote} : ${err.message}`));
-                        }
-                        else
-                        {
-                            removed.push({ file: item });
-                            done();
-                        }
+                        removed.push({ setting: item });
+                        done();
+                    }).catch((err) =>
+                    {
+                        done(new Error(`Cannot save file: ${item.remoteFilename} : ${err.message}`));
                     });
                 },
                 (err) =>
@@ -449,22 +391,22 @@ export default class Config
     }
 
     /**
-     * Get all snippet files.
+     * Get all local snippet files.
      * @param snippetsDir Snippets dir.
      */
-    private _getSnippets(snippetsDir: string): IConfig[]
+    private _getSnippets(snippetsDir: string): ISetting[]
     {
-        const results: IConfig[] = [];
+        const results: ISetting[] = [];
         try
         {
             const filenames: string[] = fs.readdirSync(snippetsDir);
             filenames.filter(junk.not).forEach((filename: string) =>
             {
-                // Add prefix `snippet-` to all snippets.
+                // Add prefix to all snippets.
                 results.push({
-                    name: filename,
-                    path: path.join(snippetsDir, filename),
-                    remote: `${Config.SNIPPET_PREFIX}${filename}`
+                    filepath: path.join(snippetsDir, filename),
+                    remoteFilename: `${VSCodeSetting.SNIPPET_PREFIX}${filename}`,
+                    type: SettingTypes.Snippets
                 });
             });
         }
@@ -476,37 +418,37 @@ export default class Config
     }
 
     /**
-     * Load file content of configs.
-     * The content will exactly be `undefined` when failure happens.
-     * @param configs VSCode configs.
-     * @param exclude Default is `true`, exclude the VSCode settings base on the exclude list of Syncing.
+     * Load the content of `VSCode Settings` files. The `content` will exactly be `undefined` when failure happens.
+     *
+     * @param settings `VSCode Settings`.
+     * @param exclude Default is `true`, exclude some `VSCode Settings` base on the exclude list of `Syncing`.
      */
-    private _loadContent(configs: IConfig[], exclude: boolean = true): Promise<IConfig[]>
+    private _loadContent(settings: ISetting[], exclude: boolean = true): Promise<ISetting[]>
     {
         return new Promise((resolve) =>
         {
             let content: string | undefined;
-            const results: IConfig[] = configs.map((item: IConfig) =>
+            const results: ISetting[] = settings.map((setting: ISetting) =>
             {
                 try
                 {
-                    if (item.name === "extensions")
+                    if (setting.type === SettingTypes.Extensions)
                     {
                         content = JSON.stringify(this._ext.getAll(), null, 4);
                     }
                     else
                     {
-                        content = fs.readFileSync(item.path, "utf8");
+                        content = fs.readFileSync(setting.filepath, "utf8");
                     }
                 }
-                catch (e)
+                catch (err)
                 {
                     content = undefined;
-                    console.error(`Syncing: Error loading config file: ${item.name}.\n${e}`);
+                    console.error(`Syncing: Error loading VSCode settings file: ${setting.type}.\n${err}`);
                 }
 
                 // Exclude settings.
-                if (exclude && item.name.includes("settings") && content)
+                if (exclude && setting.type === SettingTypes.Settings && content)
                 {
                     const settingsJSON = parse(content);
                     if (settingsJSON)
@@ -519,26 +461,27 @@ export default class Config
                     }
                 }
 
-                return { ...item, content };
+                return { ...setting, content };
             });
             resolve(results);
         });
     }
 
     /**
-     * Save item content to file or sync extensions.
-     * @param item Item of configs.
+     * Save `VSCode Setting` to file or sync extensions.
+     *
+     * @param setting `VSCode Setting`.
      */
-    private _saveItemContent(item: IConfig): Promise<ISyncStatus>
+    private _saveSetting(setting: ISetting): Promise<ISyncedItem>
     {
         return new Promise((resolve, reject) =>
         {
-            if (item.name === "extensions")
+            if (setting.type === SettingTypes.Extensions)
             {
                 try
                 {
                     // Sync extensions.
-                    const extensions: IExtension[] = parse(item.content || "[]");
+                    const extensions: IExtension[] = parse(setting.content || "[]");
                     this._ext.sync(extensions, true).then(resolve).catch(reject);
                 }
                 catch (err)
@@ -546,72 +489,56 @@ export default class Config
                     reject(new Error(`The extension list is broken: ${err.message}`));
                 }
             }
+            else if (setting.type === SettingTypes.Settings && setting.content)
+            {
+                // TODO: refactor.
+                // Merge settings.
+                let { content } = setting;
+                this._loadContent([setting], false).then((value) =>
+                {
+                    const localSettings = value[0].content;
+                    if (localSettings)
+                    {
+                        content = mergeSettings(content, localSettings);
+                    }
+
+                    // Save to disk.
+                    this._saveToFile({ ...setting, content }).then(resolve).catch(reject);
+                });
+            }
             else
             {
-                if (item.name.includes("settings") && item.content)
-                {
-                    // TODO: refactor.
-                    // Merge settings.
-                    let { content } = item;
-                    this._loadContent([item], false).then((value) =>
-                    {
-                        const localSettings = value[0].content;
-                        if (localSettings)
-                        {
-                            content = mergeSettings(content, localSettings);
-                        }
-
-                        // Save to disk.
-                        this._saveToFile({ ...item, content }).then(resolve).catch(reject);
-                    });
-                }
-                else
-                {
-                    // Save to disk.
-                    this._saveToFile(item).then(resolve).catch(reject);
-                }
+                // Save to disk.
+                this._saveToFile(setting).then(resolve).catch(reject);
             }
         });
     }
 
     /**
-     * Save the config to disk.
+     * Save the `VSCode Setting` to disk.
      */
-    private _saveToFile(config: IConfig)
+    private _saveToFile(setting: ISetting)
     {
-        return new Promise((resolve, reject) =>
-        {
-            fs.writeFile(config.path, config.content || "{}", (e) =>
-            {
-                if (e)
-                {
-                    reject(e);
-                }
-                else
-                {
-                    resolve({ file: config });
-                }
-            });
-        });
+        return fs.outputFile(setting.filepath, setting.content || "{}").then(() => ({ setting } as ISyncedItem));
     }
 
     /**
      * Check if the downloading should be continued.
      */
-    private _checkIfContinue(configs: IConfig[], saveFiles: IConfig[], removeFiles: IConfig[])
+    private _shouldContinue(settings: ISetting[], settingsToSave: ISetting[], settingsToRemove: ISetting[])
     {
         return new Promise((resolve) =>
         {
             const threshold = vscode.workspace.getConfiguration(CONFIGURATION_KEY).get<number>(CONFIGURATION_POKA_YOKE_THRESHOLD);
             if (threshold > 0)
             {
-                this._loadContent(configs, false).then((localConfigs) =>
+                this._loadContent(settings, false).then((localConfigs) =>
                 {
                     // poka-yoke - check if there have been two much changes (more than 10 changes) since the last uploading.
                     // 1. Get the excluded settings.
-                    const remoteConfigs = saveFiles.map((file) => ({ ...file }));
-                    const remoteSettings = remoteConfigs.find((item) => item.name.includes("settings"));
-                    const localSettings = localConfigs.find((item) => item.name.includes("settings"));
+                    const remoteConfigs = settingsToSave.map((setting) => ({ ...setting }));
+                    const remoteSettings = remoteConfigs.find((setting) => (setting.type === SettingTypes.Settings));
+                    const localSettings = localConfigs.find((setting) => (setting.type === SettingTypes.Settings));
                     if (remoteSettings && remoteSettings.content && localSettings && localSettings.content)
                     {
                         const localSettingsJSON = parse(localSettings.content);
@@ -625,7 +552,7 @@ export default class Config
                     }
 
                     // 2. Diff settings.
-                    const changes = this._diffSettings(localConfigs, remoteConfigs) + removeFiles.length;
+                    const changes = this._diffSettings(localConfigs, remoteConfigs) + settingsToRemove.length;
                     if (changes >= threshold)
                     {
                         const okButton = "Continue to download";
@@ -649,26 +576,26 @@ export default class Config
     }
 
     /**
-     * Calculates the number of differences between the local and remote files.
+     * Calculates the number of differences between the local and remote settings.
      */
-    private _diffSettings(localFiles: IConfig[], remoteFiles: IConfig[]): number
+    private _diffSettings(localSettings: ISetting[], remoteSettings: ISetting[]): number
     {
-        const left = this._parseToJSON(localFiles);
-        const right = this._parseToJSON(remoteFiles);
+        const left = this._parseToJSON(localSettings);
+        const right = this._parseToJSON(remoteSettings);
         return diff(left, right);
     }
 
     /**
-     * Converts the `content` of `IConfig[]` into a `JSON object`.
+     * Converts the `content` of `ISetting[]` into a `JSON object`.
      */
-    private _parseToJSON(configs: IConfig[]): any
+    private _parseToJSON(settings: ISetting[]): any
     {
         const result = {};
         let content: string;
-        for (const config of configs)
+        for (const setting of settings)
         {
-            content = config.content || "";
-            result[config.remote] = parse(content) || content;
+            content = setting.content || "";
+            result[setting.remoteFilename] = parse(content) || content;
         }
         return result;
     }